refactor(branch_pr): move branch PR lookup code into branch_pr module
- Deleted large src/branch_pr.rs and moved its contents to new
  src/branch_pr/mod.rs namespace
- Encapsulated PR-related data structures and fetch_pr_for_branch function
  within branch_pr module
- Added comprehensive tests for the branch_pr logic to src/branch_pr/tests.rs
- Minor adjustments to ref_parser module for attribute style (allow to expect)

This refactor improves module organization by isolating PR lookup logic in its
own module, improving code clarity and maintainability without functional changes.
